# Simplify the call valuer to remove the influxql.Expr references