The call valuer implementations usually just care about the interface
values. They rarely if ever cared about if it was a variable reference
or a literal. This modifies the interface to reflect that and moves the
logic for how to convert an individual AST node to an interface in the
influxql package instead so the logic is the same everywhere.