@jsternberg 2018-03-28
@Tomcat-Engineering this is what I ended up coming up with. If you think something else would be more simple to make it easier to implement functions, just tell me here so we can ensure the interface matches with the expectations of the calling developer.