@Tomcat-Engineering 2018-03-29
Is the lazy evaluation of arguments (passing a function which gives you the arg values, rather than just passing in the arg values) really worth the added complexity?  Are there common situations which this is going to speed up?  I guess the user might get an error message faster if they supply the wrong number of args or something, but that case doesn't seem worth optimising to this extent.