@Tomcat-Engineering 2018-03-29
It is possible that we might want functions to know where their arg values came from, in particular whether they were literals or not.  The example which springs to mind is unsigned vs integer literals.  The parser parses "1" as an IntegerLiteral, not an UnsignedLiteral.  The latter is only used for values which don't fit into a signed int64.  Under the proposed system, I think the math function won't be able to distinguish between a literal "1" argument and an Integer value from the database?

Imagine a hypothetical `add(x, y)` function.  If we add two unsigned values, the user might expect an unsigned value to be returned.  But if the user calls `add(x::unsigned, 1)` the math function will just see `add(unsigned, integer)` which will have to return a signed int.  

I guess the binary operator code may already have addressed this issue?