@jsternberg 2018-03-29
Here's a benchmark between the two. It seems the indirection isn't worth it from a performance standpoint like I thought it would be. I figured that it would reduce memory usage in the case of allocating 1 or 2 arguments, but it didn't.

```
benchmark           old ns/op     new ns/op     delta
BenchmarkEval-8     524           525           +0.19%

benchmark           old allocs     new allocs     delta
BenchmarkEval-8     7              7              +0.00%

benchmark           old bytes     new bytes     delta
BenchmarkEval-8     104           88            -15.38%
```

So I'll switch it to just passing in a slice of interfaces.

I would like to see a way to know the source of a value, but I don't see a way without making it more complicated. We'll likely have to get that in validation. There's some validation code in the type checker preventing binary operations between integer and unsigned if it is an unsigned literal, but I'm not really sure there's a compelling reason to have even done that. If a person is going to write that type of query, they'll just get overflow anyway so I'm not too concerned with preventing misuse of functions. The type valuer is much more complete than it was in 1.5 with this refactor so it will correctly pick up on those kinds of weird things.